+++ inferno-ml-server-types/src/Inferno/ML/Server/Client.hs
@@ -13,6 +13,7 @@ where
 
 import Data.Int (Int64)
 import Data.Proxy (Proxy (Proxy))
+import Data.UUID (UUID)
 import Inferno.ML.Server.Types
 import Servant ((:<|>) ((:<|>)))
 import Servant.Client.Streaming (ClientM, client)
@@ -23,8 +24,19 @@ statusC :: ClientM (Maybe ())
 
 -- | Run an inference parameter
 inferenceC ::
+  -- | SQL identifier of the inference parameter to be run
   Id (InferenceParam uid gid p s) ->
+  -- | Optional resolution for scripts that use e.g. @valueAt@; defaults to
+  -- 128 if not specified
   Maybe Int64 ->
+  -- | Job identifer. This is used to save execution statistics for each
+  -- inference evaluation
+  UUID ->
+  -- | Note that every item in the output stream (first element of each
+  -- outer tuple) should be declared as writable by the corresponding
+  -- inference parameter. It is the responsibility of the runtime system
+  -- (not defined in this repository) to verify this before directing
+  -- the writes to their final destination
   ClientM (WriteStream IO)
 
 -- | Cancel the existing inference job, if it exists