+++ inferno-ml-server-types/CHANGELOG.md
@@ -1,8 +1,11 @@
 # Revision History for inferno-ml-server-types
 *Note*: we use https://pvp.haskell.org/ (MAJOR.MAJOR.MINOR.PATCH)
 
-## 0.3.0
+## 0.4.0
 * Change representation of script inputs/outputs
+
+## 0.3.0
+* Add support for tracking evaluation info
 
 ## 0.2.0
 * Add `terminated` columns for DB types