+++ inferno-ml-server/CHANGELOG.md
@@ -1,5 +1,8 @@
-## 2023.5.17
+## 2023.5.27
 * Change representation of script inputs/outputs
+
+## 2023.5.22
+* Add support for tracking evaluation info
 
 ## 2023.4.3
 * Add `terminated` column to DB types